+++ lib/model/settings/defaults.dart
@@ -4,6 +4,7 @@ import 'package:aves/model/filters/recent.dart';
 import 'package:aves/model/naming_pattern.dart';
 import 'package:aves/model/settings/enums/enums.dart';
 import 'package:aves/model/source/enums/enums.dart';
+import 'package:aves/ref/mime_types.dart';
 import 'package:aves/widgets/filter_grids/albums_page.dart';
 import 'package:aves/widgets/filter_grids/countries_page.dart';
 import 'package:aves/widgets/filter_grids/tags_page.dart';
@@ -65,6 +66,7 @@ class SettingsDefaults {
   static const albumGroupFactor = AlbumChipGroupFactor.importance;
   static const albumSortFactor = ChipSortFactor.name;
   static const countrySortFactor = ChipSortFactor.name;
+  static const placeSortFactor = ChipSortFactor.name;
   static const tagSortFactor = ChipSortFactor.name;
 
   // viewer
@@ -112,6 +114,11 @@ class SettingsDefaults {
   // tag editor
 
   static const tagEditorCurrentFilterSectionExpanded = true;
+
+  // converter
+
+  static const convertMimeType = MimeTypes.jpeg;
+  static const convertWriteMetadata = true;
 
   // rendering
   static const imageBackground = EntryBackground.white;